毫秒在 32 位系统上需要使用 long long

+++ M9Dev/NSDate+M9.h
@@ -8,7 +8,7 @@
 
 #import <Foundation/Foundation.h>
 
-typedef long NSMilliseconds;
+typedef long long NSMilliseconds;
 
 extern NSMilliseconds NSMillisecondsFromTimeInterval(NSTimeInterval timeInterval);
 extern NSTimeInterval NSTimeIntervalFromMilliseconds(NSMilliseconds milliseconds);
